pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "market-sim"
version = "0.1.0"
authors = [
  { name="BerkZerker"},
]
description = "A market simulation platform with web API and live dashboard"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
    "pyjwt>=2.9.0",
    "bcrypt>=4.2.0",
    "pydantic>=2.9.0",
    "requests>=2.32.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[project.urls]
"Homepage" = "https://github.com/user/market-sim"
"Bug Tracker" = "https://github.com/user/market-sim/issues"

[tool.setuptools.packages.find]
where = ["backend"]

[project.scripts]
market-sim = "main:main"

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["backend"]
